requirements.txt
# SQLAlchemy==2.0.33
flask==3.0.3
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
flask-migrate==4.0.7
pymysql==1.1.1
pydantic==2.9.2
